function for extracting hvg
