+++ python/minicore/util/__init__.py
@@ -1,0 +1,3 @@
+from .compute_variance import variance, hvg
+
+__all__ = ["variance", "hvg"]
