+++ python/minicore/util/constants.py
